+++ beacon-db/beacon-core/src/lib.rs
@@ -6,6 +6,7 @@ pub mod file_stats;
 pub mod metrics;
 pub mod parser;
 pub mod query;
+pub mod query_executor;
 pub(crate) mod query_metrics_store;
 pub mod query_result;
 pub mod runtime;