+++ beacon-db/beacon-core/src/statement_plan/query_planner.rs
@@ -74,6 +74,7 @@ impl QueryPlanner for BeaconQueryPlanner {
                     view.name.clone(),
                     view.input.as_ref().clone(),
                     view.definition.clone(),
+                    view.or_replace,
                     session,
                 )))
             }